Eliminated volumes of compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@11562 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/csstring.cpp
@@ -64,7 +64,7 @@ void csString::ExpandIfNeeded(size_t NewSize)
       n = (NewSize + GrowBy - 1) & ~(GrowBy - 1);
     else
     {
-      n = MaxSize != 0 ? MaxSize << 1 : DEFAULT_GROW_BY;
+      n = MaxSize != 0 ? MaxSize << 1 : size_t(DEFAULT_GROW_BY);
       while (n < NewSize)
         n <<= 1;
     }
